+++ CHANGELOG.md
@@ -54,6 +54,7 @@ CHANGELOG
 * @dmlap updated to videojs-swf@4.7.1 to fix a video dimensions issue on subsequent loads ([view](https://github.com/videojs/video.js/pull/2281))
 * @mmcc added the vjs-big-play-centered class ([view](https://github.com/videojs/video.js/pull/2293))
 * @thijstriemstra added a logged error when a plugin is missing ([view](https://github.com/videojs/video.js/pull/1931))
+* @gkatsev fixed the texttrackchange event and text track display for non-native tracks ([view](https://github.com/videojs/video.js/pull/2215))
 
 --------------------